# api/openapi.yaml
openapi: 3.1.0
info:
  title: Moirai Public API
  version: 0.1.0
  description: |
    HTTP gateway for the Moirai control plane. The API uses the `loop_session`
    cookie created by the login endpoint. Mutating authenticated requests also
    require the `X-CSRF-Token` header to equal the `loop_csrf` cookie value.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/projects:
    get:
      summary: List projects
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a project
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectConfiguration'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/projects/{project_id}:
    put:
      summary: Replace a project configuration
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectConfiguration'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/projects/{project_id}/enable:
    post:
      summary: Enable scheduling for a project
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          description: Enabled project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/projects/{project_id}/disable:
    post:
      summary: Disable scheduling for a project
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          description: Disabled project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/projects/{project_id}/credentials:
    get:
      summary: Report which credentials a project has configured
      description: |
        Returns presence and timestamps only. A stored credential value is
        never returned by this API — replacing it is the only way to change it.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          description: Configured credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCredentialList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/projects/{project_id}/credentials/{kind}:
    put:
      summary: Store or replace one project credential
      description: |
        The value is encrypted at rest with the deployment's `LOOP_SECRET_KEY`.
        Requests are rejected with 422 when no key is configured, because a
        credential written without one could not be read back.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/CredentialKind'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCredentialValue'
      responses:
        '200':
          description: Configured credentials after the write
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCredentialList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: Remove one project credential
      description: |
        The project falls back to the deployment-wide credential. Removing a
        credential that is not configured succeeds.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/CredentialKind'
      responses:
        '200':
          description: Configured credentials after the removal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCredentialList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/runner-tokens:
    get:
      summary: List runner registration tokens
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Registration tokens
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunnerToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a runner registration token
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [allowedLabels]
              properties:
                allowedLabels:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Registration token; retain the token value because it is not returned by list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedRunnerToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/runner-tokens/{token_id}:
    delete:
      summary: Revoke a runner registration token
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/TokenID'
      responses:
        '204':
          description: Token revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: loop_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string }
...
    TokenID:
      name: token_id
      in: path
      required: true
      schema: { type: string }
...
    CredentialKind:
      name: kind
      in: path
      required: true
      description: >-
        `github_token`, `ssh_private_key`, or `agent:<NAME>` where `<NAME>` is
        the environment variable an agent harness reads the credential from,
        e.g. `agent:OPENROUTER_API_KEY`. Percent-encode the colon in the path.
      schema:
        type: string
        pattern: '^(github_token|ssh_private_key|agent:[A-Z_][A-Z0-9_]{0,127})$'
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or invalid
    Forbidden:
      description: A CSRF token is required for this mutation
    NotFound:
      description: Requested resource was not found
    ValidationError:
      description: The control plane rejected valid JSON input
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    ProjectConfiguration:
      type: object
      properties:
        name: { type: string }
        repositoryMode: { type: string }
        repositoryUrl: { type: string }
        localRepositoryPath: { type: string }
        defaultBranch: { type: string }
        requiredRunnerLabels:
          type: array
          items: { type: string }
        pipelineSteps:
          type: array
          items: { $ref: '#/components/schemas/PipelineStep' }
        executionImage:
          type: string
          description: Docker image used for agent and pipeline execution; empty uses runner environment.
        requireHumanApproval:
          type: boolean
          description: >-
            Opt-in human-approval gate: when true, a run stops at the
            'waiting_human' status once every automated gate (implementation,
            GitHub checks) passes, and only merges once an admin approves it
            via the decision panel. Defaults to false.
        requirePlanning:
          type: boolean
          description: >-
            Opt-in planning gate (#351): when true, the orchestrator dispatches
            a planner execution before the developer one, and its output feeds
            the developer packet's context. A run sits at the 'planning' status
            while that execution runs. Defaults to false.
    Project:
      type: object
      required: [id, name, enabled]
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        repositoryMode: { type: string }
        repositoryUrl:
          type: string
          description: >-
            Clone source for managed_clone projects. Userinfo (embedded
            credentials) is never returned; any present userinfo is replaced
            with a redacted placeholder.
        localRepositoryPath: { type: string }
        defaultBranch: { type: string }
        requiredRunnerLabels:
          type: array
          items: { type: string }
        pipelineSteps:
          type: array
          items: { $ref: '#/components/schemas/PipelineStep' }
        executionImage:
          type: string
          description: Docker image used for agent and pipeline execution; empty uses runner environment.
        requireHumanApproval:
          type: boolean
          description: >-
            Opt-in human-approval gate: when true, a run stops at the
            'waiting_human' status once every automated gate (implementation,
            GitHub checks) passes, and only merges once an admin approves it
            via the decision panel. Defaults to false.
        requirePlanning:
          type: boolean
          description: >-
            Opt-in planning gate (#351): when true, the orchestrator dispatches
            a planner execution before the developer one, and its output feeds
            the developer packet's context. A run sits at the 'planning' status
            while that execution runs. Defaults to false.
        taskSources:
          type: array
          items: { $ref: '#/components/schemas/TaskSource' }
          description: This project's configured task sources (see /api/v1/task-source-types).
    PipelineStep:
      type: object
      required: [command, timeoutSeconds, position, required]
      properties:
        command: { type: string }
        timeoutSeconds: { type: integer, minimum: 1 }
        position: { type: integer, minimum: 0 }
        required: { type: boolean }
    ProjectCredentialList:
      type: object
      required: [credentials]
      properties:
        credentials:
          type: array
          items: { $ref: '#/components/schemas/ProjectCredential' }
    ProjectCredential:
      type: object
      required: [kind, createdAt, updatedAt]
      description: >-
        A credential this project has configured. The value is deliberately
        absent: it is write-only through every layer of the system.
      properties:
        kind:
          type: string
          pattern: '^(github_token|ssh_private_key|agent:[A-Z_][A-Z0-9_]{0,127})$'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        filePath:
          type: string
          description: >-
            Where the harness reads this credential from, relative to the home
            directory the runner builds for the execution. Empty means it is
            delivered as an environment variable. A destination, not a secret,
            which is why it is reported and the value never is.
    ProjectCredentialValue:
      type: object
      required: [value]
      properties:
        value:
          type: string
          format: password
          description: >-
            The credential itself — a token, PEM private-key material, or a
            provider API key or credentials document. Must not be blank; use
            DELETE to remove a credential.
        filePath:
          type: string
          description: >-
            Only for an `agent:<NAME>` kind. Deliver the credential as a file at
            this path below the execution's home directory instead of as an
            environment variable; the variable then carries the path. Must be
            relative, with no `..` segment. Required for a subscription harness
            that reads its credentials from `~/...`, because the runner
            overrides HOME to a throwaway directory.
...
    TaskSourceSecretField:
      type: object
      required: [key, configured]
      description: Whether a secret field has a value configured -- never the value itself.
      properties:
        key: { type: string }
        configured: { type: boolean }
    TaskSource:
      type: object
      required: [id, provider, name, enabled, configuration, secrets]
      properties:
        id: { type: string }
        provider: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        configuration:
          type: object
          description: Non-secret configuration fields, shaped by this provider's descriptor. Never contains a secret field's value.
        secrets:
          type: array
          items: { $ref: '#/components/schemas/TaskSourceSecretField' }
...
    RunnerToken:
      type: object
      required: [id, allowedLabels, expiresAt]
      properties:
        id: { type: string }
        allowedLabels:
          type: array
          items: { type: string }
        expiresAt: { type: string }
        usedAt: { type: string }
        revokedAt: { type: string }
    CreatedRunnerToken:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string }
        expiresAt: { type: string }
...